Add new call StimulusModel to define the stimulus.
The new call makes it possible to define the stimulus and bring all its
informazione through the code easily. Moreover, the method
'update_stimulus' allows to update it when some parameter changes (e.g.
the magnitude diring the identification of the threshold).
Many other parts of the code have been changed according to the
introduction of this class.

Changes to be committed:
	modified:   senn.py