Fix missing module from scope. Add function to analyze solution.
timer (from timeit) and ode (from scipy.integrate) was not defined
so that every function could call them.
eqdiff function wad used not properly. Not it is passed as argumento
to each function the uses it.
A new function is defined to analyze the solution of the SENN model
called 'analyze_solution'.

 Changes to be committed:
	modified:   senn.py